pyproject.toml
[tool.poetry]
name = "vesuvius-challenge-ink-detection"
version = "0.1.0"
description = ""
authors = ["daikichiba9511 <daikichiba.tech.ml@gmail.com>"]
readme = "README.md"
# packages = [{include = "vesuvius_challenge_ink_detection"}]

[tool.poetry.dependencies]
python = "^3.9"
numpy = "^1.24.2"
pandas = "^1.5.3"
matplotlib = "^3.7.1"
seaborn = "^0.12.2"
torch = "^2.0.0"
torchvision = "^0.15.1"
opencv-python = "^4.7.0.72"
rasterio = "^1.3.6"
matplotlib-venn = "^0.11.9"
timm = "^0.6.12"
scikit-learn = "^1.2.2"
loguru = "^0.6.0"
segmentation-models-pytorch = "^0.3.2"
wandb = "^0.14.0"
albumentations = "^1.3.0"
warmup-scheduler = "^0.3"
ipdb = "^0.13.13"
monai = "^1.1.0"
ttach = "^0.0.3"
transformers = "^4.28.1"
cupy-cuda11x = "^12.0.0"

[tool.poetry.group.dev.dependencies]
pyproject-flake8 = "^6.0.0.post1"
black = "^23.1.0"
isort = "^5.12.0"
mypy = "^1.1.1"
jupyterlab = "^3.6.1"

[tool.black]
line-length = 88
target-version = ["py39"]

[tool.isort]
py_version = 39
line_length = 88
profile = "black"

[tool.flake8]
max-line-length = 88
extend-ignore = [
    "E203",
    "E266",
    "E501",
    "W503",
    "B905",
    "B907"
]
exclude = "git,.venv,__pycache__,.github,.mypy_cache,.vscode,.pytest_cache,notebooks,.ipynb_checkpoints"

[tool.mypy]
python_version = 3.9
ignore_missing_imports = true
disallow_untyped_defs = true
allow_redefinition = true
warn_redundant_casts = true
show_error_context = true
incremental = false
no_implicit_optional = true
warn_unreachable = true
show_error_codes = true
pretty = true
disallow_incomplete_defs = true
warn_no_return = true
plugins = "numpy.typing.mypy_plugin"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
